[#135] 남은 시간 계산 함수 공통으로 분리
